[#90574030] Renamed common files generator to module installer generator
