+++ lib/generators/express_admin/module_installer/USAGE
