+++ lib/generators/express_admin/module_installer/module_installer_generator.rb
@@ -1,6 +1,6 @@
 module ExpressAdmin
   module Generators
-    class CommonFilesGenerator < Rails::Generators::Base
+    class ModuleInstallerGenerator < Rails::Generators::Base
       source_root File.expand_path('../templates', __FILE__)
       attr_reader :project_name
 
@@ -9,6 +9,7 @@ module ExpressAdmin
 
         template "assets/javascripts/ajax_forms.js", File.join('app/assets/javascripts', @project_name, 'admin' ,'ajax_forms.js')
       end
+
     end
   end
 end
