+++ lib/generators/express_admin/module_installer/templates/assets/javascripts/ajax_forms.js
